Add get pet by id, fix pet pages

+++ src/PetHealthCareSystemBlazorPages/Pages/Pet/Delete.cshtml.cs
@@ -48,11 +48,11 @@ namespace PetHealthCareSystemRazorPages.Pages.Pet
                 return NotFound();
             }
 
-            //var pet = await _petService.GetPetByID(id.Value);
-            /*if (pet != null)
+            var pet = await _petService.GetPetByIdAsync(id.Value);
+            if (pet != null)
             {
                 await _petService.DeletePetAsync(pet.Id, 2002);
-            }*/
+            }
 
             return RedirectToPage("./Index");
         }
